requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.0
numpy>=1.26.0
pandas==2.1.2
openpyxl==3.1.2
googlemaps==4.10.0
python-dateutil==2.8.2
pytz==2023.3
httpx==0.24.1
